Загрузка по 5 комментов + разбиение на модули
